separate the new functionality in a component
