+++ src/app/onboarding/components/archive-creation-with-share/archive-creation-with-share.component.html
@@ -1,0 +1,27 @@
+<!-- @format  -->
+<div class="text text-share">
+  <p>
+    You’ve just taken the first step on a journey to preserving what matters the
+    most to you by creating archives that can represent yourself, your family,
+    or your community.
+  </p>
+  <div class="invitation-info">
+    <div class="img-container">
+      <img src="assets/svg/onboarding/shared_item.svg" />
+    </div>
+    <div class="sharer-info">
+      <p class="shared-item-name">{{ sharedItemName }}</p>
+      <p class="sharer-name">Shared by {{ sharerName }} on Permanent.org</p>
+    </div>
+  </div>
+  <p>
+    In order to accept {{ sharerName }}’s invitation to share their item
+    {{ sharedItemName }}, you first need to create a Permanent archive — a
+    collection of digital materials about an individual, family, group, or
+    organization — from which to access your shares.
+  </p>
+  <p>
+    Get started by designing your archive, or skip this step and create a
+    default personal archive.
+  </p>
+</div>
