+++ src/app/onboarding/components/archive-creation-with-share/archive-creation-with-share.component.scss
@@ -1,0 +1,40 @@
+/* @format */
+@import 'fonts';
+
+p {
+  color: white;
+}
+
+.text > p {
+  font-size: 16px;
+  font-family: 'UsualLight', sans-serif;
+}
+
+.invitation-info {
+  display: flex;
+}
+
+.text-share {
+  & > p {
+    margin-bottom: 1rem;
+    font-family: 'UsualRegular', sans-serif;
+  }
+
+  & > div {
+    margin-bottom: 1rem;
+  }
+}
+
+.sharer-info {
+  margin-left: 10px;
+}
+
+.shared-item-name {
+  font-family: 'UsualBold', sans-serif;
+  margin-bottom: 5px;
+}
+
+.sharer-name {
+  color: rgba(184, 187, 201, 1);
+  margin-bottom: 0px;
+}
