+++ src/app/onboarding/components/archive-creation-with-share/archive-creation-with-share.component.ts
@@ -1,0 +1,26 @@
+/* @format */
+import { Component, OnInit } from '@angular/core';
+import { ApiService } from '@shared/services/api/api.service';
+@Component({
+  selector: 'pr-archive-creation-with-share',
+  templateUrl: './archive-creation-with-share.component.html',
+  styleUrl: './archive-creation-with-share.component.scss',
+})
+export class ArchiveCreationWithShareComponent implements OnInit {
+  public hasShareToken = false;
+  public sharerName: string;
+  public sharedItemName: string;
+
+  constructor(private api: ApiService) {}
+
+  ngOnInit(): void {
+    const token = localStorage.getItem('shareToken');
+    if (token) {
+      this.api.invite.getFullShareInvite(token).then((invite) => {
+        const inviteVO = invite.getInviteVO();
+        this.sharerName = inviteVO.AccountVO.fullName;
+        this.sharedItemName = inviteVO.ArchiveVO.fullName;
+      });
+    }
+  }
+}
