+++ src/app/onboarding/components/glam/archive-creation-start-screen/archive-creation-start-screen.component.scss
@@ -68,32 +68,3 @@ p {
     gap: 32px;
   }
 }
-
-.invitation-info {
-  display: flex;
-}
-
-.text-share {
-  & > p {
-    margin-bottom: 1rem;
-    font-family: 'UsualRegular', sans-serif;
-  }
-
-  & > div {
-    margin-bottom: 1rem;
-  }
-}
-
-.sharer-info {
-  margin-left: 10px;
-}
-
-.shared-item-name {
-  font-family: 'UsualBold', sans-serif;
-  margin-bottom: 5px;
-}
-
-.sharer-name {
-  color: rgba(184, 187, 201, 1);
-  margin-bottom: 0px;
-}
